+++ src/ui/include/ui/validators/amount_line_edit.hpp
@@ -1,6 +1,7 @@
 #ifndef __UI__INCLUDE__UI__VALIDATORS__AMOUNT_LINE_EDIT_HPP__
 #define __UI__INCLUDE__UI__VALIDATORS__AMOUNT_LINE_EDIT_HPP__
 
+#include "config/finance.hpp"
 #include "line_validator.hpp"
 
 namespace ui
@@ -31,6 +32,8 @@ namespace ui
 
         void setNDecimalPlaces(std::uint8_t numberOfDecimalPlaces);
         void setOnlyPositive(bool onlyPositive);
+
+        [[nodiscard]] micro_units getAmount() const;
 
        protected:
         [[nodiscard]] QString _getRequiredErrorMessage() const override;