+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All changes and updates, that are relevant for a user will be documented here
 
 ## Next Release
 
+### Bug Fixes
+
+- Fixed: creating new profile without immediate save works now
+- Fixed: `Create Transaction` does not get the last used transaction type for creating transactions
+- Fixed: having same transaction multiple times in table view when saving after creating
+
 <!-- insertion marker -->
 ## [0.0.4](https://github.com/repo/owner/releases/tag/0.0.4) - 2026-04-18